Unify OS-independent structs
